+++ src/PackageUploader.Application.Test/CommandLineHelperIntegrationTest.cs
@@ -5,22 +5,20 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
-using PackageUploader.Application.Extensions;
 using PackageUploader.ClientApi.Client.Ingestion.TokenProvider;
 using PackageUploader.ClientApi.Client.Ingestion.TokenProvider.Config;
 using PackageUploader.ClientApi.Client.Ingestion.TokenProvider.Models;
-using System.CommandLine.Invocation;
-using System.CommandLine.Parsing;
+using System.CommandLine;
 using System.Text.Json;
 
 namespace PackageUploader.Application.Test
 {
     [TestClass]
-    public class ParameterHelperIntegrationTest
+    public class CommandLineHelperIntegrationTest
     {
         private string? _tempJsonFilePath;
         private IConfigurationRoot? _configuration;
-        private Parser? _parser;
+        private RootCommand? _rootCommand;
 
         [TestInitialize]
         public void Initialize()
@@ -45,8 +43,8 @@ namespace PackageUploader.Application.Test
 
             File.WriteAllText(_tempJsonFilePath, JsonSerializer.Serialize(configJson));
 
-            // Set up parser
-            _parser = ParameterHelper.BuildCommandLine().Build();
+            // Set up root command
+            _rootCommand = CommandLineHelper.BuildRootCommand();
         }
 
         [TestCleanup]
@@ -67,12 +65,13 @@ namespace PackageUploader.Application.Test
             var args = new[] { "GetProduct" };
 
             // Act
-            Assert.IsNotNull(_parser);
-            var result = _parser.Parse(args);
+            Assert.IsNotNull(_rootCommand);
+            var result = _rootCommand.Parse(args);
 
             // Assert
             Assert.AreEqual(1, result.Errors.Count);
-            Assert.IsTrue(result.Errors[0].Message.Contains("Option '-c' is required."));
+            Assert.IsTrue(result.Errors[0].Message.Contains("--ConfigFile") || result.Errors[0].Message.Contains("-c"),
+                $"Expected error about ConfigFile option, got: {result.Errors[0].Message}");
         }
 
         [TestMethod]
@@ -82,16 +81,15 @@ namespace PackageUploader.Application.Test
             // Arrange
             var configBuilder = new ConfigurationBuilder();
             var args = new[] { "GetProduct", "--ConfigFile", _tempJsonFilePath!, "--ProductId", "override-product-id" };
-            Assert.IsNotNull(_parser);
-            var result = _parser.Parse(args);
+            Assert.IsNotNull(_rootCommand);
+            var result = _rootCommand.Parse(args);
             
-            // Mock the invocation context to get the option value
-            InvocationContext invocationContext = new(result);
-            var configFile = invocationContext.GetOptionValue(ParameterHelper.ConfigFileOption);
-            var authMethod = invocationContext.GetOptionValue(ParameterHelper.AuthenticationMethodOption);
+            // Get option values from parse result
+            var configFile = result.GetValue(CommandLineHelper.ConfigFileOption);
+            var authMethod = result.GetValue(CommandLineHelper.AuthenticationMethodOption);
 
             // Act
-            ParameterHelper.ConfigureParameters(configFile, authMethod, configBuilder, args);
+            CommandLineHelper.ConfigureParameters(configFile, authMethod, configBuilder, args);
             _configuration = configBuilder.Build();
 
             // Assert
@@ -192,16 +190,15 @@ namespace PackageUploader.Application.Test
                 "--TenantId", "my-tenant-id" 
             };
             var configBuilder = new ConfigurationBuilder();
-            Assert.IsNotNull(_parser);
-            var result = _parser.Parse(args);
+            Assert.IsNotNull(_rootCommand);
+            var result = _rootCommand.Parse(args);
 
-            // Mock the invocation context to get the option value
-            InvocationContext invocationContext = new(result);
-            var configFile = invocationContext.GetOptionValue(ParameterHelper.ConfigFileOption);
-            var authMethod = invocationContext.GetOptionValue(ParameterHelper.AuthenticationMethodOption);
+            // Get option values from parse result
+            var configFile = result.GetValue(CommandLineHelper.ConfigFileOption);
+            var authMethod = result.GetValue(CommandLineHelper.AuthenticationMethodOption);
 
             // Act
-            ParameterHelper.ConfigureParameters(configFile, authMethod, configBuilder, args);
+            CommandLineHelper.ConfigureParameters(configFile, authMethod, configBuilder, args);
             _configuration = configBuilder.Build();
 
             // Assert
@@ -220,16 +217,15 @@ namespace PackageUploader.Application.Test
                 "--TenantId", "my-cached-tenant-id" 
             };
             var configBuilder = new ConfigurationBuilder();
-            Assert.IsNotNull(_parser);
-            var result = _parser.Parse(args);
+            Assert.IsNotNull(_rootCommand);
+            var result = _rootCommand.Parse(args);
 
-            // Mock the invocation context to get the option value
-            InvocationContext invocationContext = new(result);
-            var configFile = invocationContext.GetOptionValue(ParameterHelper.ConfigFileOption);
-            var authMethod = invocationContext.GetOptionValue(ParameterHelper.AuthenticationMethodOption);
+            // Get option values from parse result
+            var configFile = result.GetValue(CommandLineHelper.ConfigFileOption);
+            var authMethod = result.GetValue(CommandLineHelper.AuthenticationMethodOption);
 
             // Act
-            ParameterHelper.ConfigureParameters(configFile, authMethod, configBuilder, args);
+            CommandLineHelper.ConfigureParameters(configFile, authMethod, configBuilder, args);
             _configuration = configBuilder.Build();
 
             // Assert
@@ -262,16 +258,15 @@ namespace PackageUploader.Application.Test
             };
             
             var configBuilder = new ConfigurationBuilder();
-            Assert.IsNotNull(_parser);
-            var result = _parser.Parse(args);
+            Assert.IsNotNull(_rootCommand);
+            var result = _rootCommand.Parse(args);
 
-            // Mock the invocation context to get the option value
-            InvocationContext invocationContext = new(result);
-            var configFile = invocationContext.GetOptionValue(ParameterHelper.ConfigFileOption);
-            var authMethod = invocationContext.GetOptionValue(ParameterHelper.AuthenticationMethodOption);
+            // Get option values from parse result
+            var configFile = result.GetValue(CommandLineHelper.ConfigFileOption);
+            var authMethod = result.GetValue(CommandLineHelper.AuthenticationMethodOption);
 
             // Act
-            ParameterHelper.ConfigureParameters(configFile, authMethod, configBuilder, args);
+            CommandLineHelper.ConfigureParameters(configFile, authMethod, configBuilder, args);
             _configuration = configBuilder.Build();
 
             // Assert
@@ -299,16 +294,15 @@ namespace PackageUploader.Application.Test
             };
             
             var configBuilder = new ConfigurationBuilder();
-            Assert.IsNotNull(_parser);
-            var result = _parser.Parse(args);
+            Assert.IsNotNull(_rootCommand);
+            var result = _rootCommand.Parse(args);
 
-            // Mock the invocation context to get the option value
-            InvocationContext invocationContext = new(result);
-            var configFile = invocationContext.GetOptionValue(ParameterHelper.ConfigFileOption);
-            var authMethod = invocationContext.GetOptionValue(ParameterHelper.AuthenticationMethodOption);
+            // Get option values from parse result
+            var configFile = result.GetValue(CommandLineHelper.ConfigFileOption);
+            var authMethod = result.GetValue(CommandLineHelper.AuthenticationMethodOption);
 
             // Act
-            ParameterHelper.ConfigureParameters(configFile, authMethod, configBuilder, args);
+            CommandLineHelper.ConfigureParameters(configFile, authMethod, configBuilder, args);
             _configuration = configBuilder.Build();
 
             // Assert
