+++ src/PackageUploader.Application/DataOutputOptions.cs
@@ -1,0 +1,6 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace PackageUploader.Application;
+
+internal record DataOutputOptions(bool IsData);
